feat: 🎸 Display ants and nests based on server data
Huge update that moves the game simulation to the backend and also
genuinely runs player-submitted ant files for each ant. Far from the
finished product but this is now 'playable'. This is really too big for
a single commit, but I've been working very on and off working on this
project for the last 6 months so it's been difficult to keep my head
around the changes.