feat: 🎸 Don't move ants onto occupied spaces
Checks if an ant's requested movement will place them onto a nest or
another ant. If it will, the movement is ignored. For checking nests a
package is used for checking if a point is in a polygon. This is
probably overkill right now, but I like to have the flexibility with the
nest shape as it may change. One looming issue here is the order of
excution in ant actions. If two ants want to move into the same spot,
who should win? If an ant wants to move into an occupied space, but the
occupying ant also plans to move away for their action, should that be
taken into consideration? What if two ants want to swap places? For now
let's just ignore that and see if it becomes important.